+++ metaheuristicas.h
@@ -2,7 +2,6 @@
 #define METAHEURISTICAS_H_INCLUDED
 
 #include <time.h>
-
 #include "grafo.h"
 #include "lista_encadeada.h"
 
@@ -12,6 +11,6 @@ typedef struct tupla {
     int c1, c2, i1, i2;
 } tipo_tupla;
 
-int simulated_annealing(int*/*, int*/, int, grafo/*, int**/, lista*, int*, int, float, float, float, clock_t, long int);
+int simulated_annealing(int*, int, grafo, lista*, int*, int, double, double, double, clock_t, long int);
 
 #endif // METAHEURISTICAS_H_INCLUDED
